# Faster type check