# `r#type` optimization

Before:

```sh
> bench/sum/main.sh
Benchmark 1: ../../target/release/arachne < main.arc
  Time (mean ± σ):      62.2 ms ±   0.6 ms    [User: 53.6 ms, System: 6.8 ms]
  Range (min … max):    60.9 ms …  63.6 ms    46 runs

Benchmark 2: python3 main.py
  Time (mean ± σ):      36.3 ms ±   0.4 ms    [User: 32.4 ms, System: 3.1 ms]
  Range (min … max):    35.3 ms …  37.6 ms    75 runs

Summary
  python3 main.py ran
    1.72 ± 0.03 times faster than ../../target/release/arachne < main.arc
```

After:

```sh
> bench/sum/main.sh
Benchmark 1: ../../target/release/arachne < main.arc
  Time (mean ± σ):      57.2 ms ±   0.7 ms    [User: 48.7 ms, System: 6.6 ms]
  Range (min … max):    55.7 ms …  60.0 ms    50 runs

Benchmark 2: python3 main.py
  Time (mean ± σ):      36.4 ms ±   0.4 ms    [User: 32.4 ms, System: 3.2 ms]
  Range (min … max):    35.4 ms …  37.4 ms    75 runs

Summary
  python3 main.py ran
    1.57 ± 0.03 times faster than ../../target/release/arachne < main.arc
```

# `PartialEq` optimization

```sh
> bench/sum/main.sh
Benchmark 1: ../../target/release/arachne < main.arc
  Time (mean ± σ):      56.2 ms ±   0.4 ms    [User: 49.1 ms, System: 6.8 ms]
  Range (min … max):    55.4 ms …  57.1 ms    51 runs

Benchmark 2: python3 main.py
  Time (mean ± σ):      37.6 ms ±   0.5 ms    [User: 33.4 ms, System: 3.4 ms]
  Range (min … max):    36.9 ms …  39.6 ms    71 runs

Summary
  python3 main.py ran
    1.49 ± 0.02 times faster than ../../target/release/arachne < main.arc
```